Adds tests for extract_regularizer
