+++ pyunfold/tests/test_init.py
@@ -6,5 +6,5 @@ def test_version_exists():
     assert hasattr(pyunfold, '__version__')
 
 
-def test_version_exists():
+def test_iterative_unfold_exists():
     assert hasattr(pyunfold, 'iterative_unfold')
